This is runnable via a simple nginx container
If updating the index.html page, you'll need to update the block in
nginx.conf that has the contents of index.html dropped into it, so that
nginx can serve the compiled application.

To run the container, the environment variables API_KEY and STOPS_LIST
must be provided so they are available in the html served up.
API_KEY is a trimet developer key
STOPS_LIST is a comma-separated list of trimet stops